+++ src/webview/components/memory-widget.tsx
@@ -19,9 +19,10 @@ import { WebviewIdMessageParticipant } from 'vscode-messenger-common';
 import { Memory } from '../../common/memory';
 import { WebviewSelection } from '../../common/messaging';
 import { MemoryOptions, ReadMemoryArguments, SessionContext } from '../../common/messaging';
+import { MemoryDisplayConfiguration } from '../../common/webview-configuration';
 import { ColumnStatus } from '../columns/column-contribution-service';
 import { HoverService } from '../hovers/hover-service';
-import { Decoration, MemoryDisplayConfiguration, MemoryState } from '../utils/view-types';
+import { Decoration, MemoryState } from '../utils/view-types';
 import { createAppVscodeContext, VscodeContext } from '../utils/vscode-contexts';
 import { MemoryTable } from './memory-table';
 import { OptionsWidget } from './options-widget';
