+++ src/webview/hovers/hover-service.tsx
@@ -17,8 +17,9 @@
 import * as React from 'react';
 import { HOST_EXTENSION } from 'vscode-messenger-common';
 import { logMessageType } from '../../common/messaging';
+import { MemoryDisplayConfiguration } from '../../common/webview-configuration';
 import { MemoryAppState } from '../memory-webview-view';
-import { Disposable, MemoryDisplayConfiguration } from '../utils/view-types';
+import { Disposable } from '../utils/view-types';
 import { messenger } from '../view-messenger';
 
 export interface HoverableDetails {
